requirements.txt
redis
redis-py-cluster
requests
aiohttp
jsonpath
PyYAML
